// frontend/src/app/api/model/experiment-response.ts
/**
 * Experiment Tracking Platform API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TagResponse } from './tag-response';

export interface ExperimentResponse {
  id?: string;
  projectId?: string;
  name?: string;
  description?: string;
  createdAt?: string;
  tags?: Array<TagResponse>;
}